Moving prototype to its own folder.  Small text update for clarity.
